+++ comfy/comfy_types/__init__.py
@@ -1,7 +1,6 @@
 import torch
 from typing import Callable, Protocol, TypedDict, Optional, List
 from .node_typing import IO, InputTypeDict, ComfyNodeABC, CheckLazyMixin, FileLocator
-from .input_types import ImageInput, AudioInput, VideoInput, VideoFromFile, VideoFromComponents, VideoContainer, VideoCodec, VideoComponents
 
 
 class UnetApplyFunction(Protocol):
@@ -44,12 +43,4 @@ __all__ = [
     ComfyNodeABC.__name__,
     CheckLazyMixin.__name__,
     FileLocator.__name__,
-    ImageInput.__name__,
-    AudioInput.__name__,
-    VideoInput.__name__,
-    VideoFromFile.__name__,
-    VideoFromComponents.__name__,
-    VideoContainer.__name__,
-    VideoCodec.__name__,
-    VideoComponents.__name__,
 ]
